+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - allow configuring the program used to `open` associated files
 - prompt user before actually quitting CoBib (#33)
+- implements the Search command (#7, !12)
 
 ### Fixed
 - the default value of the `open` command was not set correctly